+++ packages/terra-core-docs/src/terra-dev-site/test/text/TextItalics.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Text from 'terra-text';
+
+const TextItalics = () => (
+  <div>
+    <Text id="textItalic" isItalic>Italics</Text>
+  </div>
+);
+
+export default TextItalics;