+++ packages/terra-table/tests/wdio/table-spec.js
@@ -1,38 +1,38 @@
 Terra.describeViewports('Table', ['medium'], () => {
   it('displays table with Fill', () => {
-    browser.url('/raw/tests/terra-table/table/fill');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/fill');
     Terra.validates.element('fill');
   });
   it('displays table with HeaderCheckMark', () => {
-    browser.url('/raw/tests/terra-table/table/header-check-mark');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/header-check-mark');
     Terra.validates.element('header check mark');
   });
   it('displays table with HeaderFooter', () => {
-    browser.url('/raw/tests/terra-table/table/header-footer');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/header-footer');
     Terra.validates.element('header footer');
   });
   it('displays table with Padding', () => {
-    browser.url('/raw/tests/terra-table/table/padding');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/padding');
     Terra.validates.element('padding');
   });
   it('displays Row', () => {
-    browser.url('/raw/tests/terra-table/table/row');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/row');
     Terra.validates.element('row');
   });
   it('displays Section', () => {
-    browser.url('/raw/tests/terra-table/table/section');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/section');
     Terra.validates.element('section');
   });
   it('displays Striped table', () => {
-    browser.url('/raw/tests/terra-table/table/striped');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/striped');
     Terra.validates.element('striped');
   });
   it('displays table with Custom Cell Content', () => {
-    browser.url('/raw/tests/terra-table/table/custom-cell');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/custom-cell');
     Terra.validates.element('custom cell content');
   });
   it('displays a Multi Demo table', () => {
-    browser.url('/raw/tests/terra-table/table/multi-demo');
+    browser.url('/raw/tests/cerner-terra-core-docs/table/multi-demo');
     Terra.validates.element('multi demo');
   });
 });